news_main/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
python-dotenv==1.0.0
djangorestframework-simplejwt==5.2.2
